requirements.txt
torch>=2.5.0
torchvision
numpy<2
opencv-python
timm
scipy
scikit-image
kornia
einops

tqdm
prettytable
tabulate

huggingface-hub>0.25
accelerate
transformers

requests
python-dotenv
supabase
runpod
gdown
uuid